v1

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: tzorai <tzorai@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 10:02:33 by tzorai            #+#    #+#             */
-/*   Updated: 2023/11/11 17:07:31 by tzorai           ###   ########.fr       */
+/*   Updated: 2023/11/28 16:34:03 by tzorai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	int	i;
 
 	i = ft_strlen(s);
 	while (i >= 0)
